+++ js/explorer.js
@@ -2,28 +2,34 @@
 
 var EXP = {};
 
-requirejs([
+require([
+	'lib/data',
 	'lib/sound',
 	'lib/controller',
 	'lib/effects',
 	'lib/util',
 	'lib/ui',
 	'lib/physics',
+	'lib/objects',
 	'lib/items',
 	'lib/planet',
+	'lib/space',
 	'lib/camera',
 	'lib/player',
 	'lib/engine',
 	'lib/debug'
-], function(Sound, Controller, Effects, Util, Ui, Physics, Items, Planet, Camera, Player, Engine, Debug) {
+], function(Data, Sound, Controller, Effects, Util, Ui, Physics, Objects, Items, Planet, Space, Camera, Player, Engine, Debug) {
 
 	EXP = {
+		data: Data,
 		sound: Sound,
 		planet: Planet,
+		space: Space,
 		player: Player,
 		engine: Engine,
 		camera: Camera,
 		controller: Controller,
+		objects: Objects,
 		items: Items,
 		effects: Effects,
 		util: Util,
@@ -41,20 +47,23 @@ requirejs([
         	camera: EXP.camera
         });
         
-        EXP.planet.create();
+        var planet = EXP.planet.create();
         console.log(EXP.planet);	
         
         var player = EXP.player.create({controllable: true});
         EXP.engine.add(player);
         EXP.camera.follow(player);
         
-        var health = EXP.items.health.create({x: 8*32, y: 6*32});
+        EXP.engine.add( EXP.items.create(EXP.data.items.gold));
+        EXP.engine.add( EXP.items.create(EXP.data.items.rock));
+        
+        var health = EXP.objects.health.create({x: 8*32, y: 6*32});
         EXP.engine.add(health);
         
-        var health2 = EXP.items.health.create({x: -2*32, y: -4*32, boost: 50});
+        var health2 = EXP.objects.health.create({x: -2*32, y: -4*32, boost: 50});
         EXP.engine.add(health2);
         
-        var portal = EXP.items.portal.create({x: 2*32, y: 2*32});
+        var portal = EXP.objects.portal.create({x: 2*32, y: 2*32});
         EXP.engine.add(portal);
         
         EXP.engine.run();