+++ js/lib/data.js
@@ -1,0 +1,12 @@
+define([
+	'lib/data/items'
+],function (Items) {
+
+	var Data = {
+		items: Items
+	};
+
+		
+	return Data;
+	
+});