+++ js/lib/debug.js
@@ -2,8 +2,16 @@ define(function () {
 
 	var Debug = {
 	    
-	    showCollider: false
+	    colliders: false
 	
+	};
+	
+	Debug.showColliders = function () {
+		this.colliders = true;
+	};
+	
+	Debug.hideColliders = function () {
+		this.colliders = false;
 	};
 		
 	return Debug;