+++ js/lib/min/debug-min.js
@@ -1,1 +1,1 @@
-define(function(){var e={showCollider:!1};return e});
+define(function(){var i={colliders:!1};return i.showColliders=function(){this.colliders=!0},i.hideColliders=function(){this.colliders=!1},i});