+++ js/lib/objects.js
@@ -1,0 +1,15 @@
+define([
+	'lib/objects/health',
+	'lib/objects/portal'
+], function (Health, Portal) {
+
+	var Objects = {
+	    
+	    health: Health,
+	    portal: Portal
+	
+	};
+		
+	return Objects;
+    
+});