+++ js/lib/ui.js
@@ -1,14 +1,16 @@
 define([
 	'lib/ui/healthbar',
 	'lib/ui/shieldbar',
-	'lib/ui/jetfuelbar'
-], function (Healthbar, Shieldbar, Jetfuelbar) {
+	'lib/ui/jetfuelbar',
+	'lib/ui/inventorybar'
+], function (Healthbar, Shieldbar, Jetfuelbar, Inventorybar) {
 
 	var Ui = {
 	    
 	    healthbar: Healthbar,
 	    shieldbar: Shieldbar,
-	    jetfuelbar: Jetfuelbar
+	    jetfuelbar: Jetfuelbar,
+	    inventorybar: Inventorybar
 	
 	};
 	
@@ -17,6 +19,7 @@ define([
 		this.shieldbar.update();
 		this.healthbar.update();
 		this.jetfuelbar.update();
+		this.inventorybar.update();
 	
 	};
 		
@@ -25,6 +28,7 @@ define([
 		this.shieldbar.render();
 		this.healthbar.render();
 		this.jetfuelbar.render();
+		this.inventorybar.render();
 		
 	};