+++ js/lib/ui/healthbar.js
@@ -4,7 +4,7 @@ define(function () {
 		percent: 1,
 		target: 1,
 		x: 20,
-		y: 315,
+		y: 24+16*2,
 		icon: new Image()
 	};
 	
@@ -20,14 +20,14 @@ define(function () {
 	
 		EXP.engine.ctx.drawImage(this.icon, 0, 0, 24, 24, this.x, this.y, 24, 24);
 		
-		EXP.engine.ctx.fillStyle = '#000';
-		EXP.engine.ctx.fillRect(this.x+24+9, this.y + 8, 102, 8);
+		EXP.engine.ctx.fillStyle = 'rgba(0,0,0,0.5)';
+		EXP.engine.ctx.fillRect(this.x+20, this.y + 4, 103, 16);
 		
-		EXP.engine.ctx.fillStyle = '#444';
-		EXP.engine.ctx.fillRect(this.x+24+10, this.y + 9, 100, 6);
+		EXP.engine.ctx.fillStyle = 'rgba(255,255,255,0.3)';
+		EXP.engine.ctx.fillRect(this.x+20, this.y + 8, 100, 8);
 			
-		EXP.engine.ctx.fillStyle = '#74e494';
-		EXP.engine.ctx.fillRect(this.x + 24 + 10, this.y + 9, this.percent * 100, 6);
+		EXP.engine.ctx.fillStyle = '#29fcb5';
+		EXP.engine.ctx.fillRect(this.x + 20, this.y + 8, Math.round(this.percent * 100), 8);
 	
 	};